Removed the wakeup_interval/2 method that made the timed_supervisor always restart it's child after 1 hour if the next scheduled start was more than an hour away.
This only applies to when a child is terminated and but the timed_supervisor has been configured to reschedule on termination. The default branch (master) will always reschedule it after an hour if we're still within the scheduler timeframe. This is best explained with a case:

If I have a process running and it's configured to run between 10 - 14 but I terminate it at 12 and want it to restart at 10 tomorrow. The default branch of the timed supervisor will restart the process in 3600 seconds. This branch will wait untill the 10 tomorrow.

I'm working at way to merge this into the master without breaking current implementations; probably by adding an option.